Squashed commit of the following: Branch: refactor/reduceDuplicateCode
commit cde1fb1a557d5ada5b73c4e7d07fccb41e4343e7
Author: Nicholas Winder <ntwinder@gmail.com>
Date:   Fri Mar 3 11:48:42 2017 +0900

    There is now a abstract Class which handles most of the preference set up.  All the preferences share this code

+++ library/src/test/java/com/nfx/android/preferencehelper/preferences/LongPreferenceTest.java
@@ -50,7 +50,7 @@ public class LongPreferenceTest {
     public void testIsSet() throws Exception {
         assertThat("Should return false due to preference no being set",
                 longPreference.isSet(), is(false));
-        longPreference.set(0);
+        longPreference.set((long) 0);
         assertThat("Should return true as preference has just been set",
                 longPreference.isSet(), is(true));
     }
@@ -65,7 +65,7 @@ public class LongPreferenceTest {
 
     @Test
     public void testDelete() throws Exception {
-        longPreference.set(0);
+        longPreference.set((long) 0);
         assertThat("Should return true as preference has just been set",
                 longPreference.isSet(), is(true));
         longPreference.delete();
